requirements.txt
streamlit
langchain-openai
langchain-community
langchain-chroma
python-dotenv
